# packages/contracts/openapi/midas.openapi.yaml
openapi: 3.1.0
info:
  title: Midas Marketplace API
  version: 0.1.0
  description: >-
    Contrato executável dos cortes de identidade, SellerAccount e finanças:
    capabilities fail-closed, confirmação de pagamento por webhook, resolução
    administrativa auditável, saldo do seller e payout manual com segregação.
servers:
  - url: http://localhost:3001
paths:
...
  /v1/auth/register:
    post:
      operationId: registerUser
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterUser' }
      responses:
        '202':
          description: Cadastro pendente de verificação.
          content:
            application/json:
              schema:
                type: object
                required: [registrationStatus]
                properties:
                  registrationStatus: { const: PENDING_VERIFICATION }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /v1/auth/email-verifications:
    post:
      operationId: verifyUserEmail
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: { type: string, minLength: 32, maxLength: 256 }
      responses:
        '204': { description: E-mail verificado. }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/auth/login:
    post:
      operationId: createSession
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 320 }
                password: { type: string, minLength: 1, maxLength: 256 }
      responses:
        '204':
          description: Sessão criada em cookie HttpOnly.
          headers:
            Set-Cookie: { schema: { type: string } }
        '401': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /v1/auth/session:
    delete:
      operationId: deleteCurrentSession
      tags: [Identity]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204': { description: Sessão atual revogada ou já ausente. }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/me/sessions:
    get:
      operationId: listCurrentUserSessions
      tags: [Identity]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Sessões próprias ativas.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionList' }
        '401': { $ref: '#/components/responses/Problem' }
  /v1/me/sessions/{sessionId}:
    delete:
      operationId: revokeCurrentUserSession
      tags: [Identity]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204': { description: Sessão própria revogada ou já ausente. }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/seller-accounts:
    post:
      operationId: createSellerAccount
      tags: [Sellers]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSellerAccount' }
      responses:
        '201':
          description: Conta e membership OWNER criadas atomicamente.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SellerAccount' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/me/seller-accounts:
    get:
      operationId: listCurrentUserSellerAccounts
      tags: [Sellers]
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Contas vinculadas à sessão.
          content:
            application/json:
              schema:
                type: object
                required: [data, asOf]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SellerAccount' } }
                  asOf: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Problem' }
  /v1/seller-accounts/{sellerAccountId}/members:
    get:
      operationId: listSellerAccountMembers
      tags: [Sellers]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: sellerAccountId
          in: path
          required: true
          schema: { type: string, pattern: '^sac_[0-9a-f-]{36}$' }
      responses:
        '200':
          description: Members autorizados sem e-mail ou contato.
          content:
            application/json:
              schema:
                type: object
                required: [data, asOf]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SellerMember' } }
                  asOf: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /v1/provider-webhooks/{providerCode}:
    post:
      operationId: receivePaymentProviderWebhook
      tags: [ProviderWebhooks]
      security: []
      description: >-
        Recebe bytes originais do webhook e delega validação de assinatura e
        parsing ao adapter registrado. Não confirma pagamento sem provider
        disponível e evento verificável.
      parameters:
        - name: providerCode
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
      requestBody:
        required: true
        content:
          '*/*':
            schema: {}
      responses:
        '202':
          description: Evento autenticado recebido, inclusive replay idempotente.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderWebhookAck' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: midas_session
  parameters:
    CsrfHeader:
      name: X-Midas-CSRF
      in: header
      required: true
      schema: { const: '1' }
...
  responses:
    Problem:
      description: Erro tipado.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    RegisterUser:
      type: object
      additionalProperties: false
      required: [email, password, displayName, acceptedTermsVersion]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 256, writeOnly: true }
        displayName: { type: string, minLength: 2, maxLength: 100 }
        acceptedTermsVersion: { type: string, minLength: 1, maxLength: 64 }
    CreateSellerAccount:
      type: object
      additionalProperties: false
      required: [displayName, accountType]
      properties:
        displayName: { type: string, minLength: 2, maxLength: 120 }
        accountType: { enum: [INDIVIDUAL, ORGANIZATION] }
    SellerAccount:
      type: object
      additionalProperties: false
      required: [sellerAccountId, displayName, accountType, sellerAccountStatus, membershipRole, version]
      properties:
        sellerAccountId: { type: string, pattern: '^sac_[0-9a-f-]{36}$' }
        displayName: { type: string }
        accountType: { enum: [INDIVIDUAL, ORGANIZATION] }
        sellerAccountStatus: { enum: [ONBOARDING_REQUIRED, ACTIVE, SUSPENDED] }
        membershipRole: { enum: [OWNER, MANAGER, OPERATOR, FINANCE_VIEWER] }
        version: { type: integer, minimum: 1 }
    SellerMember:
      type: object
      additionalProperties: false
      required: [sellerMembershipId, userId, displayName, membershipRole, membershipStatus, validFrom, validUntil, version]
      properties:
        sellerMembershipId: { type: string, pattern: '^smb_[0-9a-f-]{36}$' }
        userId: { type: string, pattern: '^usr_[0-9a-f-]{36}$' }
        displayName: { type: string }
        membershipRole: { enum: [OWNER, MANAGER, OPERATOR, FINANCE_VIEWER] }
        membershipStatus: { enum: [ACTIVE, REVOKED, EXPIRED] }
        validFrom: { type: string, format: date-time }
        validUntil: { type: [string, 'null'], format: date-time }
        version: { type: integer, minimum: 1 }
    SessionList:
      type: object
      additionalProperties: false
      required: [data, asOf]
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [sessionId, createdAt, expiresAt, current]
            properties:
              sessionId: { type: string, format: uuid }
              createdAt: { type: string, format: date-time }
              expiresAt: { type: string, format: date-time }
              current: { type: boolean }
        asOf: { type: string, format: date-time }
    FinanceUuid:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
...
    ProviderWebhookAck:
      type: object
      additionalProperties: false
      required: [duplicate, processingResult, paymentId]
      properties:
        duplicate: { type: boolean }
        processingResult:
          enum:
            - SETTLED
            - PAYMENT_QUARANTINED
            - IGNORED_PENDING
            - IGNORED_FAILED
            - PAYMENT_NOT_FOUND
        paymentId:
          oneOf:
            - $ref: '#/components/schemas/FinanceUuid'
            - type: 'null'
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, detail, correlationId, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }